conda/requirements.txt
docker
requests
PyYAML
jinja2
